Changing quantity back to Int for both the request and the response DTO

+++ Sources/GroceryAppSharedDTO/GroceryItemDTO.swift
@@ -10,9 +10,9 @@ import Foundation
 public struct GroceryItemRequestDTO: Codable {
     public var title: String
     public var price: Decimal
-    public var quantity: Int64
+    public var quantity: Int
     
-    public init(title: String, price: Decimal, quantity: Int64) {
+    public init(title: String, price: Decimal, quantity: Int) {
         self.title = title
         self.price = price
         self.quantity = quantity
@@ -23,9 +23,9 @@ public struct GroceryItemResponseDTO: Codable {
     public var id: UUID
     public var title: String
     public var price: Decimal
-    public var quantity: Int64
+    public var quantity: Int
 
-    public init(id: UUID, title: String, price: Decimal, quantity: Int64) {
+    public init(id: UUID, title: String, price: Decimal, quantity: Int) {
         self.id = id
         self.title = title
         self.price = price
